Move C API to separate directory
