+++ src/c-api/c_molecular_weights.cxx
@@ -13,9 +13,10 @@
 #include <gauxc/molecular_weights.hpp>
 #include <gauxc/load_balancer.h>
 #include <gauxc/load_balancer.hpp>
-#include <gauxc/util/c_molecular_weights.hpp>
-#include <gauxc/util/c_load_balancer.hpp>
-#include <gauxc/util/c_status.hpp>
+
+#include "c_molecular_weights.hpp"
+#include "c_load_balancer.hpp"
+#include "c_status.hpp"
 
 namespace GauXC::C {
 extern "C" {
@@ -26,6 +27,7 @@ GauXCMolecularWeightsFactory gauxc_molecular_weights_factory_new(
   const char* kernel_name,
   const GauXCMolecularWeightsSettings settings
 ) {
+  detail::gauxc_status_init(status);
   GauXCMolecularWeightsFactory mwf{};
   mwf.hdr = GauXCHeader{GauXC_Type_MolecularWeightsFactory};
   mwf.ptr = nullptr;
@@ -36,10 +38,8 @@ GauXCMolecularWeightsFactory gauxc_molecular_weights_factory_new(
       std::string(kernel_name),
       detail::convert_molecular_weights_settings( settings )
     );
-    status->code = 0;
   } catch (std::exception& e) {
-    status->code = 1;
-    status->message = detail::strdup(e.what());
+    detail::gauxc_status_handle(status, 1, e.what());
   }
   return mwf;
 }
@@ -48,6 +48,7 @@ GauXCMolecularWeights gauxc_molecular_weights_factory_get_instance(
   GauXCStatus* status,
   const GauXCMolecularWeightsFactory mwf
 ) {
+  detail::gauxc_status_init(status);
   GauXCMolecularWeights mw{};
   mw.hdr = GauXCHeader{GauXC_Type_MolecularWeights};
   mw.ptr = nullptr;
@@ -56,10 +57,8 @@ GauXCMolecularWeights gauxc_molecular_weights_factory_get_instance(
   try {
     MolecularWeights mw_instance = detail::get_molecular_weights_factory_ptr(mwf)->get_instance();
     mw.ptr = new MolecularWeights( std::move(mw_instance) );
-    status->code = 0;
   } catch (std::exception& e) {
-    status->code = 1;
-    status->message = detail::strdup(e.what());
+    detail::gauxc_status_handle(status, 1, e.what());
   }
   return mw;
 }
@@ -68,6 +67,7 @@ GauXCMolecularWeights gauxc_molecular_weights_factory_get_shared_instance(
   GauXCStatus* status,
   const GauXCMolecularWeightsFactory mwf
 ) {
+  detail::gauxc_status_init(status);
   GauXCMolecularWeights mw{};
   mw.hdr = GauXCHeader{GauXC_Type_MolecularWeights};
   mw.ptr = nullptr;
@@ -76,10 +76,8 @@ GauXCMolecularWeights gauxc_molecular_weights_factory_get_shared_instance(
   try {
     auto mw_instance_ptr = detail::get_molecular_weights_factory_ptr(mwf)->get_shared_instance();
     mw.ptr = new std::shared_ptr<MolecularWeights>( std::move(mw_instance_ptr) );
-    status->code = 0;
   } catch (std::exception& e) {
-    status->code = 1;
-    status->message = detail::strdup(e.what());
+    detail::gauxc_status_handle(status, 1, e.what());
   }
   return mw;
 }
@@ -89,6 +87,7 @@ void gauxc_molecular_weights_modify_weights(
   const GauXCMolecularWeights mw,
   const GauXCLoadBalancer lb
 ) {
+  detail::gauxc_status_init(status);
   try {
     if (mw.owned) {
       if (lb.owned)
@@ -109,10 +108,8 @@ void gauxc_molecular_weights_modify_weights(
           **detail::get_load_balancer_shared(lb)
         );
     }
-    status->code = 0;
   } catch (std::exception& e) {
-    status->code = 1;
-    status->message = detail::strdup(e.what());
+    detail::gauxc_status_handle(status, 1, e.what());
   }
 }
 
@@ -121,7 +118,7 @@ void gauxc_molecular_weights_delete(
   GauXCStatus* status,
   GauXCMolecularWeights* mw
 ) {
-  status->code = 0;
+  detail::gauxc_status_init(status);
   if(mw == nullptr) return;
   if(mw->ptr != nullptr) {
     if (mw->owned)
@@ -137,7 +134,7 @@ void gauxc_molecular_weights_factory_delete(
   GauXCStatus* status,
   GauXCMolecularWeightsFactory* mwf
 ) {
-  status->code = 0;
+  detail::gauxc_status_init(status);
   if(mwf == nullptr) return;
   if(mwf->ptr != nullptr)
     delete detail::get_molecular_weights_factory_ptr(*mwf);
