+++ src/c-api/c_molecule.cxx
@@ -13,29 +13,30 @@
 #include <gauxc/atom.hpp>
 #include <gauxc/molecule.h>
 #include <gauxc/molecule.hpp>
-#include <gauxc/util/c_molecule.hpp>
-#include <gauxc/util/c_status.hpp>
+
+#include "c_molecule.hpp"
+#include "c_status.hpp"
 
 namespace GauXC::C {
 
 extern "C" {
 
 GauXCMolecule gauxc_molecule_new(GauXCStatus* status) {
+  detail::gauxc_status_init(status);
   GauXCMolecule mol{};
   mol.hdr = GauXCHeader{GauXC_Type_Molecule};
   mol.ptr = nullptr;
 
   try {
     mol.ptr = new Molecule();
-    status->code = 0;
   } catch (std::exception& e) {
-    status->code = 1;
-    status->message = detail::strdup(e.what());
+    detail::gauxc_status_handle(status, 1, e.what());
   }
   return mol;
 }
 
 GauXCMolecule gauxc_molecule_new_from_atoms(GauXCStatus* status, GauXCAtom* atoms, size_t natoms) {
+  detail::gauxc_status_init(status);
   GauXCMolecule mol{};
   mol.hdr = GauXCHeader{GauXC_Type_Molecule};
   mol.ptr = nullptr;
@@ -48,17 +49,15 @@ GauXCMolecule gauxc_molecule_new_from_atoms(GauXCStatus* status, GauXCAtom* atom
       mol_ptr->push_back(detail::convert_atom(atoms[i]));
     }
     mol.ptr = mol_ptr;
-    status->code = 0;
   } catch (std::exception& e) {
     delete mol_ptr;
-    status->code = 1;
-    status->message = detail::strdup(e.what());
+    detail::gauxc_status_handle(status, 1, e.what());
   }
   return mol;
 }
 
 void gauxc_molecule_delete(GauXCStatus* status, GauXCMolecule* mol) {
-  status->code = 0;
+  detail::gauxc_status_init(status);
   if (mol == nullptr) return;
   if (mol->ptr != nullptr)
     delete detail::get_molecule_ptr(*mol);
@@ -66,7 +65,7 @@ void gauxc_molecule_delete(GauXCStatus* status, GauXCMolecule* mol) {
 }
 
 size_t gauxc_molecule_natoms(GauXCStatus* status, const GauXCMolecule mol) {
-  status->code = 0;
+  detail::gauxc_status_init(status);
   if (mol.ptr == nullptr) return 0;
   return detail::get_molecule_ptr(mol)->natoms();
 }
@@ -76,7 +75,7 @@ bool gauxc_molecule_equal(
   const GauXCMolecule mol_a,
   const GauXCMolecule mol_b
 ) {
-  status->code = 0;
+  detail::gauxc_status_init(status);
   if (mol_a.ptr == nullptr || mol_b.ptr == nullptr) return false;
   return *detail::get_molecule_ptr(mol_a) == *detail::get_molecule_ptr(mol_b);
 }
