+++ src/c-api/c_molgrid.cxx
@@ -14,9 +14,10 @@
 #include <gauxc/molgrid.h>
 #include <gauxc/molgrid.hpp>
 #include <gauxc/molgrid/defaults.hpp>
-#include <gauxc/util/c_molecule.hpp>
-#include <gauxc/util/c_molgrid.hpp>
-#include <gauxc/util/c_status.hpp>
+
+#include "c_molecule.hpp"
+#include "c_molgrid.hpp"
+#include "c_status.hpp"
 
 namespace GauXC::C {
 extern "C" {
@@ -29,6 +30,7 @@ GauXCMolGrid gauxc_molgrid_new_default(
   const enum GauXC_RadialQuad radial_quad,
   const enum GauXC_AtomicGridSizeDefault grid_size
 ) {
+  detail::gauxc_status_init(status);
   GauXCMolGrid mg{};
   mg.hdr = GauXCHeader{GauXC_Type_MolGrid};
   mg.ptr = nullptr;
@@ -42,16 +44,14 @@ GauXCMolGrid gauxc_molgrid_new_default(
       static_cast<AtomicGridSizeDefault>(grid_size)
     );
     mg.ptr = new MolGrid(grid_map);
-    status->code = 0;
   } catch (std::exception& e) {
-    status->code = 1;
-    status->message = detail::strdup(e.what());
+    detail::gauxc_status_handle(status, 1, e.what());
   }
   return mg;
 }
 
 void gauxc_molgrid_delete(GauXCStatus* status, GauXCMolGrid* mg) {
-  status->code = 0;
+  detail::gauxc_status_init(status);
   if (mg == nullptr) return;
   if (mg->ptr != nullptr)
     delete detail::get_molgrid_ptr(*mg);
