+++ src/c-api/c_types.cxx
@@ -22,11 +22,13 @@
 #include <gauxc/xc_integrator.h>
 #include <gauxc/matrix.h>
 
+#include "c_status.hpp"
+
 namespace GauXC::C {
 extern "C" {
 
 void gauxc_object_delete(GauXCStatus* status, void** obj) {
-   status->code = 0;
+   detail::gauxc_status_init(status);
    if(obj == nullptr) return;
 
    struct GauXCObject {
@@ -97,7 +99,7 @@ void gauxc_object_delete(GauXCStatus* status, void** obj) {
        break;
      }
      default: {
-       status->code = 1; // Unknown type
+       detail::gauxc_status_handle(status, 1, "Unknown object type in gauxc_object_delete");
        break;
      }
    }
@@ -108,10 +110,10 @@ void gauxc_objects_delete(
   void** ptrs,
   size_t nptrs
 ) {
-   status->code = 0;
+   detail::gauxc_status_init(status);
    for(void** ptr = ptrs; ptr < ptrs + nptrs; ++ptr) {
       if(*ptr != nullptr) {
-         gauxc_object_delete( status, ptr );
+         gauxc_object_delete(status, ptr);
          if(status->code != 0) return;
       }
    }
