+++ src/external/c_hdf5_read.cxx
@@ -11,10 +11,11 @@
  */
 #include <gauxc/external/hdf5.h>
 #include <gauxc/external/hdf5.hpp>
-#include <gauxc/util/c_status.hpp>
-#include <gauxc/util/c_molecule.hpp>
-#include <gauxc/util/c_basisset.hpp>
-#include <gauxc/util/c_matrix.hpp>
+
+#include "c_status.hpp"
+#include "c_molecule.hpp"
+#include "c_basisset.hpp"
+#include "c_matrix.hpp"
 #include "hdf5_util.hpp"
 
 namespace GauXC::C {
@@ -33,12 +34,11 @@ void gauxc_molecule_read_hdf5_record(
   const char* fname,
   const char* dset
 ) {
-  status->code = 0;
+  detail::gauxc_status_init(status);
   try {
     read_hdf5_record( *detail::get_molecule_ptr(mol), std::string(fname), std::string(dset) );
   } catch(std::exception& e) {
-    status->code = 1;
-    status->message = detail::strdup(e.what());
+    detail::gauxc_status_handle(status, 1, e.what());
   }
 }
 
@@ -55,12 +55,11 @@ void gauxc_basisset_read_hdf5_record(
   const char* fname,
   const char* dset
 ) {
-  status->code = 0;
+  detail::gauxc_status_init(status);
   try {
     read_hdf5_record( *detail::get_basisset_ptr(basis), std::string(fname), std::string(dset) );
   } catch(std::exception& e) {
-    status->code = 1;
-    status->message = detail::strdup(e.what());
+    detail::gauxc_status_handle(status, 1, e.what());
   }
 }
 
@@ -77,7 +76,7 @@ void gauxc_matrix_read_hdf5_record(
   const char* fname,
   const char* dset
 ) {
-  status->code = 0;
+  detail::gauxc_status_init(status);
   detail::CMatrix& mat = *detail::get_matrix_ptr(matrix);
   try {
     HighFive::File file( std::string(fname), HighFive::File::ReadOnly );
@@ -89,8 +88,7 @@ void gauxc_matrix_read_hdf5_record(
 
     dataset.read(mat.data());
   } catch(std::exception& e) {
-    status->code = 1;
-    status->message = detail::strdup(e.what());
+    detail::gauxc_status_handle(status, 1, e.what());
   }
 }
